refactor: Differentiate between cmd and args
We use the `exec` action helper to run our docker login command. The
`exec` function has the following signature:

```
 * @param     commandLine        command
 * @param     args               optional arguments
 * @param     options            optional exec options
```

Right now we specify some arguments as part of the `commandLine` (`login`)
while we pass the other arguments (e.g. username) as part of the `args`
array.

This commit now passes all arguments as an array.

This matches the signature of [`spawn`][1] and clarifies that this will
not be executed as a shellcommand and rather directly without the
involvement of a shell. This is critical because we don't need to think
about escaping special characters that might exist in the username and
password (only for windows the exec helper actually has to run this
through a `cmd` and does a hard job to properly escape everything)

[1]: https://github.com/actions/toolkit/blob/daf8bb00606d37ee2431d9b1596b88513dcf9c59/packages/exec/src/toolrunner.ts#L443